Vr core basic hand functionality 1
Created scripts:
HandController - Contains get-properties for the controllers input. (Added IsTriggerDown, IsTriggerStateUp, IsTriggerStateDown action bool)
HandFunctionality - Contains the functionality for what the controllers are supposed to be able to do. (Added Funtionality to be able to pick up gameobjects whith the PickupAble script)
PickupAble - Makes the object pickupable by the controllers ()
SnapToTransform - Makes the object snap back to a transform when not held by the hands (not implemented yet)