Volume slider and scoreboard changes
The volume sliders now change static floats for the respective volume types.

The scoreboard now changes both with difficulty select, and level select. This means that we now have a way of knowing what level is selected. There is no real feedback to the player that they have selected a particular level though, but that will be worked on after this